requirements.txt
praisonaiagents[llm]
python-dotenv
requests
beautifulsoup4
mcp
